requirements.txt
attrs==21.2.0
certifi==2018.11.29
cffi==1.15.0
click==8.0.3
coverage==6.0.2
cryptography==35.0.0
dnspython==1.16.0
Flask==2.0.2
Flask-Cors==3.0.10
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
packaging==21.0
pluggy==1.0.0
py==1.10.0
pycairo==1.20.1
pycparser==2.20
PyGObject==3.40.1
PyJWT==2.3.0
pymongo==3.12.1
pyparsing==3.0.0
pytest==6.2.5
PyYAML==6.0
six==1.16.0
toml==0.10.2
Werkzeug==2.0.2
